requirements.txt
requests>=2.28.0
python-docx>=1.0.0
